techpack: audio: q6voice: Force AVCS CVP version 1 on legacy SoC
Force the AVCS version to CVP_VERSION_1 for all of the AVCS
services on sdm636: It cannot execute the Q6 Service Version
command on it's Q6 DSP because of older firmware not supporting it.
We know that for this SoC the version is always 1, hence
forcing it is totally safe.

Signed-off-by: sohamxda7 <sensoham135@gmail.com>
Signed-off-by: Oktapra Amtono <oktapra.amtono@gmail.com>
Signed-off-by: clarencelol <clarencekuiek@icloud.com>
Signed-off-by: Anush02198 <Anush.4376@gmail.com>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: HeroBuxx <me@herobuxx.me>