arm64: dts: sdm660: Remove useless memory dump region
This reserved memory dump region is intended to be used with the memory dump v2 driver, but we've disabled that (@ Tashar02: We're going to be disabling that with the defconfig imports) and we don't need this memory dumping functionality. Remove the unused region and disable the associated driver node to save memory.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: HeroBuxx <me@herobuxx.me>